Update patch set 2
Patch Set 2: Code-Review-1

> (1 comment)
 > 
 > I don't like this change. The code used to fit in a single
 > screenful, now it is separated in three separate files.
 > 
 > Note that I have only scored it -1, not -2. If you are seeing a
 > value to this change for end users, CDT developers or CDT
 > extenders, then you are free to submit it. If this code was in a
 > part of CDT we were not actively trying to move away from, that may
 > also have changed my score and the amount of time I spend reviewing
 > it.

I agree with Jonah, to me this code is harder to understand as you have to jump in and out of multiple files for what is <20 lines of code.

If there were some extensibility benefit then I could see an reason to justify it but at the moment it just seems to make our code base more complex not less.

Patch-set: 2
Label: Code-Review=-1