Bug 562498: Prevent end of life Arduino from installing on newer CDT
When a user does an upgrade with Arduino installed and an upgrade is done
now that Arduino is removed from master they can get into a broken state
because Arduino plug-ins had not traditionally defined upper ranges.

This commit adds such limits so users will get a fail to upgrade error to
CDT 10 if they have Arduino installed.

Change-Id: I05e7f40e21dd1a2f511d3d0ced13dac49ec0bec1