WIP Bug 540737: Dispose SWT Color objects
With support for 24-bit color we can have huge numbers of different
colors and unlike ealier versions with 16 or so colors we can't
stick the 16M colors in the color registry in good conscience because
the memory use would be unreasonable.

However with this patch, disposing colors as we go leads to a measureable
performance loss (about 7.5Mb/s vs 9Mb/s).

Therefore I need a better way to dispose of the colors. Or simply not
bother disposing of them as SWT at the moment does not actually
allocate any resources on new Color() anyway. Indeed the patchset
Ief4ccaf1adb70967301411b012c0474608f425cb removed the creation of
tons of colors that were never being disposed.

Change-Id: I8fca0a9b4a729e041c227d4a5a5da92edc239a8e