Bug 562697 - Fix comparison of qualifier type and pointer type
CDT has two representations for a pointer type wrapped in a
qualifier type: it can be an IPointerType with some qualifiers
set on it directly, or an IPointerType wrapped in an
IQualifierType. (We prefer the first representation to avoid
creating too many wrappers, but sometimes the second one arises
during template instantiation.) This patch makes sure that two
such types can compare equal even if they use different
representations.

Change-Id: Ia8c7d227c74378aae74f04545b9a69103c14e74b