Update patch set 4
Patch Set 4:

@William you are right, the Stream API and Optional are not debugger friendly - they are rather friendly for cleaner control flow expression. 
My point is simple: if debug is required that means the code needs rework. I believe that unit tests should do "debug" work for us. I tried to make a step to this direction with "patch set 1" - to extract code to a smaller types that will allow easier test coverage. And you also rejected it :)

@Jonah actually this is serios discussion far beyond the scope of this ticket. 
I do not see any sense to upgrade BREE to Java 11 (breaking change) and suppress code upgrade to Java 8 at the same time. May be we should even degrade BREE to Java 5 for a number of bundles/packages and make them a kind of LTS. This is what we have in EMF that is extra-stable and can still work with Eclipse 3.x

Patch-set: 4