Update patch set 4
Patch Set 4:

> > @William you are right, the Stream API and Optional are not
 > > debugger friendly - they are rather friendly for cleaner control
 > > flow expression.
 > > My point is simple: if debug is required that means the code
 > needs
 > > rework. I believe that unit tests should do "debug" work for us.
 > I
 > > tried to make a step to this direction with "patch set 1" - to
 > > extract code to a smaller types that will allow easier test
 > > coverage. And you also rejected it :)
 > >
 > 
 > Once you are making code harder to read, understand or debug to
 > achieve smaller types I think that is taking the aim too far. To be
 > honest I find this entire change unnecessary, but I won't block it.

Well, "harder to read, understand and debug" is subjective criteria. For me the big hierarchies are harder to read and maintain.

However I can agree that change like this is better to deliever in a bigger scope together with unit tests, to not have mixed style in one method/class. From this point of view Stream and Optional usage should be either maximized or completely eliminated in order to maintain a procedural style that is so convenient for spending our lives in the debugger.

Patch-set: 4