Update patch set 1
Abandoned

We can revisit at some point, but basically the API of colors in SWT was written for a limited palette display, which is not the case anymore so disposing of colors in practice is not a problem. It seems that writing a bunch of code to manage the color resources would be a waste of time.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon