Update patch set 4
Patch Set 4:

(8 comments)

I think it is great to get some tests for the importing and exporting of memory and other such fairly complicated business logic. So everything here is a step forward over having no tests. I have done a quick review and shared my thoughts FWIW. I think the tests themselves are still lacking. Because the tests only test import against export, it means that if the "same" bug existed in the import as export then the tests would not catch that at all. It seems that the importer test should check that emulated memory has the expected values, and no other writes and that the emulated memory should be preloaded with known values and then check the export is indeed correct.

Here are a few silly bugs that the tests wouldn't catch by my read:

- if the offsets did not work at all, i.e. the importer always started at address 0 and the exporter did too, this test would work.
- the importer padded out writes so that 0 was written at the end of the import range, the exported values would be correct, but the memory would be corrupted
- the importer added 1 to each byte and the exporter subtracted one from each byte

Those examples are silly, but show some extra stuff that can be added to the tests. In addition I noted in EmulateMemory that the memory emulation appeared to be rely on the memory pattern of the import. For example you could use import and export, but offset the export start address by a byte?

Back to the main purpose of this bug, the performance aspect. The Emulate memory can include such a test. For example, the srec export test can check that no more than X number of memory reads happened. That would prevent this particular issue from regressing in the future.

Patch-set: 4