Bug 558536: Revert "Bug 549013 - Eclipse menu won't display on Ubuntu 16.04 LTS when SSH/Serial command shells are opened"
Using the mouse and not the focus listeners leads to lots
of weird, unintended consequences. For example:

- If mouse does not exit the control (because, e.g. it was not
in the control) then captureKeyEvents(false) is never issued
- If mouse does leave control, then control stops capturing
events. This means that although the cursor and focus is still
on the control, keyboard shortcuts no longer go to the control,
so Ctrl-C becomes copy again. This means, depending on where you
move the mouse changes the behaviour of the control.

This reverts commit 19351cbc2bf009659e6269cb6f1f7a1fd720d317.

Change-Id: I4f57c659e21d823df049b095159a34e5c110ef29
(cherry picked from commit 56d741a50c715d20bde1df020661edf8c85f9430)