Update patch set 1
Patch Set 1:

Not really elegant code but I needed a feedback if it's the right fix from semantic point of view. I'd like to receive Nathan's feedback. Then I can improve the code style. The example reported in the bug report now it works but I want to see if I introduced a regression looking at test report when it will be ready.

Patch-set: 1