Update patch set 2
Patch Set 2:

(2 comments)

This is a tricky area - there may very well be other solutions to this problem, and now that we don't need to support Java 8 at the same time (purpose of this bug) there may be other options. Some of the history of what was tried is in Bug 542644.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0