Update patch set 1
Patch Set 1:

Martin, for your review. These are two of the few failing tests fixed. Not sure the evolution of the code, presumably at one point the code didn't differentiate between system and user includes? But that is just a guess as to why these were failing.

It may be the correct fix is to add user includes too so that both methods are tested fully.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0