Update patch set 1
Patch Set 1: Code-Review+1

> Martin, for your review. These are two of the few failing tests
 > fixed. Not sure the evolution of the code, presumably at one point
 > the code didn't differentiate between system and user includes? But

It does differenciate. But the compiler output does contain system-includes, so thanks for this.

 > It may be the correct fix is to add user includes too so that both
 > methods are tested fully.

I don't think we need to account for user (local) includes here.

The GccOutputProcessor is responsible for parsing the output of gcc/clang/nvcc when it is invoked to spit out it built-ins.
No -I commandline option from the compile_commnds.json is passed through to gcc when doing built-in detection, so no need to care for local includes. (In contrast --sysroot is passed:)

Patch-set: 1
Reviewer: Gerrit User 276327 <276327@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1