Update patch set 1
Patch Set 1:

I supplied an aggregator pom 'cmake/aggregator/pom.xml' to hook it in to the build.
Having a 'cmake/pom.xml' suggests it is a parent pom (and is referenced as a parent in the module poms, but isn't).
Anyway, a *real* parent pom would be useful, it provides a way to override some of the nastiness (mainly surefire config) inherited from the CDt parent for all modules here.

Patch-set: 1