Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Also it looks like a change is needed in the manifest as this change does not build:
> 
> 02:41:11  [ERROR] /home/jenkins/agent/workspace/cdt-verify-test-cdt-ui-only-pipeline/launch/org.eclipse.cdt.docker.launcher/src/org/eclipse/cdt/internal/docker/launcher/ContainerDataVolumeDialog.java:[63] 
> 02:41:11  [ERROR] 	import org.mandas.docker.client.messages.ContainerMount;
> 02:41:11  [ERROR] 	       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
> 02:41:11  [ERROR] Access restriction: The type 'ContainerMount' is not API (restriction on classpath entry '/home/jenkins/.m2/repository/p2/osgi/bundle/org.mandas.docker-client/3.2.1.v20200519-1937/org.mandas.docker-client-3.2.1.v20200519-1937.jar')

This should be handled differently.  All docker stuff should be wrapped by Linux Tools so the com.spotify and org.mandas stuff is hidden.  I'll try and make the change after our status meeting.  What is the API change that is requiring a 5.0.0?  This should be hidden as well either by adding a new interface to add new functionality which is ok in a minor release.  Let me know and I'll fix it ASAP.

Patch-set: 1