Update patch set 1
Patch Set 1:

Sorry for the broken submission.  I wanted to get it posted last night in order to get CDT Bot to check the build for me, but I was not able to get back to it until now.  I'll fix the failed build and adjust to match Jeff's changes.  I need to ask Jeff some questions in another review before I can do that, but will update with fixes ASAP.

Patch-set: 1
Label: Verified=0