Bug 522289: Handle case of no terminal connectors available
The existing code had the assumption that >= 1 connector was installed
in the IDE. It is difficult to configure the IDE with no connectors as
the base terminal feature includes connectors, but it is technically
possible. So in that case display a simple error message and avoid
a hidden error (and stack trace to log).

Change-Id: Id17a1e5bff9aec251aa6f7fe6ff4e96cd2a01564