Update patch set 1
Patch Set 1:

(8 comments)

Thanks Jantje - this looks good. The items that are left are fairly trivial and mostly stylistic changes I don't want to spend too much time on.

One subtle bug is to consider is my comment in Env.java on line 54.

In addition, this change needs to be documented in API breakages html.

Patch-set: 1
Label: Code-Review=0