Update patch set 4
Patch Set 4:

> @Michael The size of change does not matter while the potential
 > impact does. In this particular case I can see that the size of
 > returned array now can be less than the size of input array. Do we
 > assume somewhere that size should match like we do inside the
 > modified function? The test with data that finally activates the
 > introduced Objects::nonNull predicate can demonstrate this fix in
 > action.
Don't get me wrong, I'm not questioning unit tests. What I tried to say is that I can't figure a way to synthesize a simple unit test that would exhibit the behaviour. It just so happens in my code.

Again, this fix does what is done throughout the whole CDT code with ArrayUtils.trim. Do a search for your own peace of mind, this is CDT style! Using streams was Marco's idea ...

Sorry, guys but please understand that this is now taking too much of my time.

Patch-set: 4