Update patch set 1
Patch Set 1:

Just to explain: for me returning `null` elements in array (or collection) is bad idea, but I do not know at the moment the right reaction on `null` value _before_ putting it to the result array. Silent filtering will fix this NPE but will hide the underlying problem.

Ypur solution LGTM as it allow to fix the NPE without affecting others. The root cause can be discussed using another Gerrit.

Patch-set: 1