Update patch set 4
Patch Set 4:

> I don't know how I can create a simple isolated test case that triggers the problem. I hope in the light of such simple fix the rules can be bent ....

OK - I am fine with someone familiar with the code bending that rule. I can become familiar with that code, but I am certainly not familiar with it at the moment so my ability to quickly review it is affected by the lack of the test. I apologize for not making that clear when I provided my review. 

I do appreciate the fix. The above is probably moot as Marco (who is familiar with the code and seems to have his computer working again) has proposed an alternative fix.

As it is too late for 2020-06 / CDT 9.11.1, this fix can be backported to 9.11 branch and we can release it as soon as we want as CDT 9.11.2. Otherwise a fix will wait until 2020-09 / CDT 10.0.0 to be released/widely available.

> However we really need to remove all array utils stuff replacing with something more Java updated style like:
> 
> Arrays.stream(result).filter(Objects::nonNull).toArray(ICPPFunction[]::new)

FWIW - I disagree that Arrays.stream(result).f... is better than the ArrayUtils.trim().

The one method call of ArrayUtils.trim() is much easier to read code than the streaming.

But more importantly, the switch from ArrayUtils.trim to the Arrays.stream has two key different behaviours:
1) ArrayUtils.trim() does not make a copy of the array if there are no nulls, it simply returns the input after a check there are no nulls.
2) trim() only trims the end of the array for nulls. The Arrays.stream version removes null anywhere in the array. This is equivalent to the ArrayUtils.removeNulls() 

Finally, ArrayUtils.trim() is a finely crafted piece of code that should have good runtime performance behaviour. In the case of no nulls, only one element of the array needs to be checked and no allocations are needed. In the case of some nulls, a binary search is done to find the first null and a single allocation is made for the new array.

Note: Using Java 8 streams to implement some of ArrayUtils may be a good idea, but not for trim based on the above.

Patch-set: 4
Label: Code-Review=0