Update patch set 2
Patch Set 2:

My thinking on this is:

It's fine for a private function like CPPSemantics.selectByArgumentCount() to return an array containing nulls, if there is an efficiency reason for it. (In this case, the caller performs further filtering of the functions, so removing nulls in selectByArgumentCount() would mean an extra allocation).

However, a public interface like ICPPDeferredFunction.getCandidates() should not return an array containing nulls.

So, I think it would be appropriate to call ArrayUtil.removeNulls() at the call site of CPPDeferredFunction.createForCandidates() in CPPSemantics.resolveFunction() (or perhaps in the implementation of createForCandidates()), and then checks in callers of getCandidates() like this one would not be necessary.

Patch-set: 2
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0