Update patch set 4
Patch Set 4:

Martin, you may have already had this in hand, but with this patch I can see the markers that you are creating are not showing in the UI. See https://git.eclipse.org/r/#/c/164839/ for a possible fix. Feel free to include that in a future patch, or let me know if you want me to rebase on master and submit it.

Patch-set: 4
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0