Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

> As I tried to describe at cdt-dev: my point is to have CompileCommandsJsonParser functionality available not for CMake only and go to more common bundles with proposed name "o.e.c.jsoncdb*"
> Having marker super-type looks like a good compromise if you think that CompileCommandsJsonParser has something really specific to cmake.

My understanding from previous conversations is that while cmake uses compile_commands.json as well as other tools, there are lots of subtle differences between them.

I think this is a good place to focus on getting this in (i.e. increasing our velocity) - if (and when) someone else wants to add support for an additional tool that uses compile_commands.json this code can be refactored as appropriate based on those needs. To keep things as "malleable" as possible we can continue to limit the internal implementation to non-API so that our hands are not tied.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+2