Update patch set 1
Patch Set 1:

"copyright corrected" sounds strange, as it looks right before the change. I would use 
"Switched from <> to <>" for comment to describe the change more clearly.

For header update please add "Bug 562003" to describe the origin of change, there is a lot of examples of this in the code base.

Patch-set: 1
Label: Code-Review=0