Update patch set 9
Patch Set 9:

Hi Martin - this looks like great progress on this. 2 of the three sets of verify jobs have the same number of tests before and after and the third one is pretty close.

Once you merge this to master a quick check on https://ci.eclipse.org/cdt/job/cdt-master/ to make sure that the number of tests hasn't changed unexpectedly would be great too.

Unless you need a further review from me, go ahead and submit when it is ready.

Patch-set: 9