Update patch set 3
Patch Set 3:

(2 comments)

Thanks for this fix. As you can see I provided an update that I would like you to do a review of.

You may also notice that there was a todo in the code I was able to delete! Thanks!

This change breaks API, but the timing is good as we have a major release coming up. If we didn't, then we would have to jump through some hoops to make sure that the binary API contract was maintained.

Patch-set: 3
Label: Verified=0