Update patch set 9
Patch Set 9:

HI Marin,

How about compare each of the verify logs. But it is easiest if you simply look at HTML test result. It will say total number of tests and +/- compared to last run. 

The only problem with that method is it compares to the previous Run (build #-1), not to some golden version. 

Let me know if you want me to make a comparison as a second set of eyes on it. 

HTH, 
Jonah

Patch-set: 9