Update patch set 2
Patch Set 2:

I'm not sure if this code did actually compile on Windows before or if GCC is more picky than VS.
While going over the code and trying to get it to compile using mingw64, I noticed that there are a lot of potential memory leaks (malloc but no free etc). I also noticed that there is a cast from "HANDLE" to "int" for stdin/stdout/stderr, but on 64-bit Windows, the "HANDLE" type is 64-bit and "int" is only 32-bit!

As I see it, there are 2 possible ways to handle these issues:
1. Fix the leaks in the C code and do extensive testing...
2. Start using JNA when possible.
  a. The Windows registry can be access just fine with JNA, no need for a custom library to do this.
  b. listtasks.exe can more or less be replaced by an implementation using JNA too. The only thing that does not appear to be available in JNA is to list 16-bit applications. Since 16-bit applications are only supported on non 64-bit Windows versions, I think it's safe to drop this requirement as Eclipse only supports 64-bit JRE on Windows now.
  c. The spawner.dll and starter.exe is another story. These need to have the leaks fixed in C code.

The benefit of using JNA compared to these JNI implementations is that the logic is implemented in java rather than C code and can more easily be debugged and is less likely to crash the JVM.

Looking forward to some comments on this one!

Patch-set: 2