Update patch set 9
Patch Set 9:

> HI Marin,
 > 
 > How about compare each of the verify logs. But it is easiest if you
 > simply look at HTML test result. It will say total number of tests
 > and +/- compared to last run.
 > 
 > The only problem with that method is it compares to the previous
 > Run (build #-1), not to some golden version.

Well, you are the project lead; trading gold for clay:-)

Sorry, I am not familar enough with the differences in the jenkins master-build job and the verification build to even find a _copper_ solution here.

Want I want to proof before merging is:
After this gerrit got merged, will the master-build job run with less tests than before?
Will it introduce new failing tests? And if so, which of the new failing tests was never run in a test suite (in order to deal with them)?
I'm fine with more tests being run with this change, if they do not fail.

I will have to prove again when I remove the surefire include patterns in some of the poms that override it before resolving the issue.

 > Let me know if you want me to make a comparison as a second set of
 > eyes on it.

Of course I want.

Patch-set: 9