Update patch set 2
Patch Set 2:

> (1 comment)

 > I really like the idea of using JNA where possible. The barrier to
 > debug is much lower that way. I assume JNA either wasn't around or
 > was not popular when this code was originally written. I don't have
 > experience with it, but I have used equivalent in Python in the
 > past and I appreciated it there.
 > 
 > So a big +1 on JNA. Note that Orbit at the moment has JNA 4.5.1
 > available (https://download.eclipse.org/tools/orbit/downloads/drops/R20200529191137/)
 > but there is a 5.5.0 release (https://search.maven.org/artifact/net.java.dev.jna/jna/5.5.0/jar)
 > now. If you think you may
 > want 5.5.0 let me know and I can get it into Orbit sooner rather
 > than later.

I don't remember what the big difference is between 4.5.x and 5.5.x, but as it's a major version number and this is "new" development, I suggest that we go for the 5.5 version if possible to skip the need for at least one upgrade in the future. Do you agree?

 > Thank you for tackling this part of CDT. I started to look at how
 > to fix the build up, but didn't get far.
 > 
 > One thing, I am working on upgrading WinPTY version we ship, so
 > don't spend too much time on that particular area. See Bug 562776.

Noted! As there were no reference to WinPTY from the makefiles or the bat files, I thought that this was a manual process and decided to push that update to a later commit. In any case, I wont touch it!

Patch-set: 2
Reviewer: Gerrit User 5071 <5071@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0