Update patch set 5
Patch Set 5:

(1 comment)

> @Torbjörn thank you for the contribution!

Thanks for the review and feedback @Alexander!

> Although the idea to reuse com.sun.jna looks good to me I don't think that extension point is the right way to implement it. As we do not have platform support to declare "singleton extension point" or "internal extension point" - this will open a door to have unwanted extension. And I already can see the code in this Gerrit that tries to emulate OSGi platform filtering.
> 
> Moreover, the idea of having cross-platform WindowsRegistry API looks questionable to me. I know that it was introduced far before your Gerrit, but may be it is good time to change it.
> 
> IMO the WindowsRegistry API should be available for .win32 fragment only
> or
> we can go SWT way and have a type like `SystemRegistry` implemented by each os-specific fragment by its own way.

I agree with that this should generally be kept in the win32 fragment. The reason why I did not go this route is that the org.eclipse.cdt.utils.WindowsRegistry class exists since before and I did not want to break the API for the entire world unless I had to. And just to be fair, the filtering was present before to, so I did more or less the same thing as before, just that before it was filtering based on if it could load a library that only the win32 fragment(s) were providing.
In any case, I think that the redesign is somewhat out of the scope. The topic here is to move away from the JNI. If you think that we should also clean up the API, I think that should be a separate ticket and I'm not sure if that should be done before or after switching from JNI to JNA.

Looking forward to your views on the cleanup @Alexander and @Jonah.

Patch-set: 5
Reviewer: Gerrit User 5071 <5071@97ee7c02-f12f-4043-b43e-dea463d88b31>