Update patch set 5
Patch Set 5:

> Patch Set 4:
> IMO the WindowsRegistry API should be available for .win32 fragment only

I took the liberty to just try to move the WindowsRegistry class from the org.eclipse.cdt.core plugin to the org.eclipse.cdt.core.win32 fragment. After the move, all hell broke loose. It appears that the assumption that there is class named WindowsRegistry and that it returns null for all non-Windows platforms is spread all over the CDT code base. The calls to the WindowsRegistry.getRegistry() is done in code that is not just for win32, but for all platforms.
Moving the class to a win32 fragment means that all those places will also need to be split in a win32 fragment and the "common" code.

My trivial move ended up breaking the following plugin/fragments that will likely need some redesign:
* org.eclipse.cdt.core
* org.eclipse.cdt.managedbuilder.gnu.ui
* org.eclipse.cdt.msw.build
* org.eclipse.cdt.qt.core
* org.eclipse.cdt.core.tests
* org.eclipse.cdt.build.gcc.core

> or
> we can go SWT way and have a type like `SystemRegistry` implemented by each os-specific fragment by its own way.

I don't know how this interface should look like to be a non-Windows centered one but still have enough details to retrieve the correct resource in the Windows registry.

My vote for the moment is to accept this extension point solution, although not an optimal one. If you have any ideas, feel free to try them out @Alexander.

Patch-set: 5
Reviewer: Gerrit User 5071 <5071@97ee7c02-f12f-4043-b43e-dea463d88b31>