Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Does this work for both unicode and non-unicode processes? I remember there were some problems with that. There might be a bugzilla about that somewhere too..

Thanks for the review and the feedback!
It should work for both. JNA will automatically select ASCII or UNICODE versions of the Windows API function depending on the "w32.ascii" system property. According to my tests on Eclipse 2020-06 on a Win7 64-bit system, I always appears to be using the UNICODE version of the API.

I tried for find any old bug entry in bugzilla regarding the ASCII/UNICODE topic, but I couldn't find any.
What I did find was https://bugs.eclipse.org/bugs/show_bug.cgi?id=236330, that you resolved back in the day. During the work to add win32-x86_64 support, the define UNICODE was introduced and forced for all builds of listtasks.exe.
Is it good enough that JNA automatically selects version or do we need to force it to UNICODE always? I think it would be possible to force this by sub-classing and overriding the load of the library, but it would be quite a bit more code to have in the CDT project. It would also mean that all the utility classes in JNA would be forbidden to use since they would end up using the default version of the API rather than the version that we would force... I'm not sure that it's a good way to go for try to force UNICODE always.

Patch-set: 1