Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > Hmm, I don't thing that executable flag exists on Windows so I don't think that commit changes anything. Unless I'm wrong?
> > 
> > Well, they sort of does. If you clone the git repository in cygwin, the files will have 644 as permission and when you then try to run/load them in Eclipse, you will get permission denied. Setting the executable flag in Cygwin fixes this and so should also cloning with the executable flag set do.
> 
> I don't get how Cygwin environment is affecting Eclipse, so I'll just clear my vote

I created a simple hello world application and placed the executable in c:\tmp.
The following is executed in Cygwin on a Win7 system:

 azoff@win7 /cygdrive/c/tmp
 $ chmod 755 helloworld.exe; ls -l helloworld.exe; cmd /c helloworld.exe; chmod 644 helloworld.exe; ls -l helloworld.exe; cmd /c helloworld.exe
 -rwxr-xr-x+ 1 azoff None 77450 Jul  5 08:52 helloworld.exe*
 Hello World!
 -rw-r--r--+ 1 azoff None 77450 Jul  5 08:52 helloworld.exe
 Access is denied.

As you can see, the permission change does have an effect.
So, with this said, do you agree that running 'git clone ...' would also have different outcome for a file that is under version control if it has executable flag set, or not set, in git?

There is no connection between the Cygwin environment and Eclipse, but there is a connection between the modes on a file on the file system and Eclipse. I can get the same behavior without Cygwin if I right click on the file in the explorer and select Properties, Security, Edit, click on my user and uncheck "Read & Execute".

Patch-set: 1