Update patch set 1
Patch Set 1:

> My git clone has all the executable flags by default. Yours doesn't, why?
> I have a setup that's as old as when CDT changed to Git (10 years ago?) and one recent, both have all executable flags. Looking into more, there is the core.fileMode config, which is set to false for me in all cases. Maybe the default changed recently. Or maybe the default is different depending on the git client. Can you confirm that yours is set to true?
> 
> $ git config --get core.fileMode
> false
> 
> I normally install the official "Git for Windows" and then use git bash.

Yes, I can confirm that the value for core.fileMode is "true".
"Git for Windows" is (AFAIK) a fork and is not using the official Git source tree. To my knowledge, Git in Cygwin is as close to the official version as you can get on Windows.

Patch-set: 1