Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > My git clone has all the executable flags by default. Yours doesn't, why?
> > I have a setup that's as old as when CDT changed to Git (10 years ago?) and one recent, both have all executable flags. Looking into more, there is the core.fileMode config, which is set to false for me in all cases. Maybe the default changed recently. Or maybe the default is different depending on the git client. Can you confirm that yours is set to true?
> > 
> > $ git config --get core.fileMode
> > false
> > 
> > I normally install the official "Git for Windows" and then use git bash.
> 
> Yes, I can confirm that the value for core.fileMode is "true".
> "Git for Windows" is (AFAIK) a fork and is not using the official Git source tree. To my knowledge, Git in Cygwin is as close to the official version as you can get on Windows.

It's a weird situation where the official download on the git website is Git for Windows but might not be the closest to the source tree. I can confirm that even with the latest version core.fileMode is still set to false. Anyway, could you put in the commit message this fix is in the context where core.fileMode is set to "true", I think it's useful information.
Also you might be worth to check other binaries such as the ones in the equinox launcher and swt.

Patch-set: 1
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0