Update patch set 1
Patch Set 1:

> It's a weird situation where the official download on the git website is Git for Windows but might not be the closest to the source tree. I can confirm that even with the latest version core.fileMode is still set to false. Anyway, could you put in the commit message this fix is in the context where core.fileMode is set to "true", I think it's useful information.

Well, both yes and no. From what I've understood, the official git tree does not compile on Windows natively and thus, the "friendly fork" was born to fix that. For some reason, the code is not in a state where it could be merged with the official tree. This is what I've understood when just googling on this topic for a bit, but I can be completely off base here 😊
Since Cygwin has a so good compatibility layer for Windows, I guess they can use the official tree and still have it working.
I'll update the commit message to mention core.fileMode=true.

> Also you might be worth to check other binaries such as the ones in the equinox launcher and swt.
I do not have those in my workspace, but I guess they could potentially have the same issue.

Patch-set: 1