Update patch set 2
Patch Set 2:

Hi! I remember while doing the first 64 bit version of the fragment on Windows, it was not possible to  use MinGW without depending on extra MinGW dlls. I had to use depends.exe to double check and also test on a fresh machine. In the end, I used Windows SDK (hence the nmake usage) and was able to not depend on additional msvcr.dll (so no need for users to install Microsoft Visual C++ Redistributable). I also saw Windows SDK as a more reproducible environment as MinGW was in a very unclear and heterogeneous state and it was unclear how long someone could reproduce a similar environment.
All that to say, to be extra careful with introducing dependencies and having a reproducible environment.

Patch-set: 2
CC: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>