Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Ed Merks does not bless the changes to .setup with anything except the dedicated editor

Yes and I don't bless generating the .target from .setup since we'd depend on Oomph for the build. But I think I'd be ready to change my opinion if we can determine that switching back to a non-generated .target would be easy/low-risk if something goes wrong (Oomph bug or Oomph goes away).

Patch-set: 1