Update patch set 1
Patch Set 1:

Of course I understand the trade-offs and concerns.   But personally I despise editing a *.target file (why does it start download artifacts right away?) and I resent that I have to wait so long to resolve a TP with hundreds of duplicates in each workspace for things I have downloaded already.  But then I have a dozen specialized installations and workspaces, because I work on, and contribute to, many projects.  If I had only one project to worry about, I guess I wouldn't be as concerned.

Patch-set: 1