Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2: Code-Review+2
> 
> LGTM, will wait a bit before submitting

PS As what you are removing was not API I don't think you need to wait any longer (nor link it to bug 561389, nor document it in the removals). Of course some non-official CDT API is essentially API so documenting its removal is a good thing.

Patch-set: 2
CC: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>