Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (2 comments)
> 
> Sorry for coming back to this so late - it turns out I wrote the above comments a while ago, but never submitted them. The recent gerrit upgrade shows gerrits with unsubmitted draft comments so I was able to see them now.

Good to see some activity on this review, but part of the original fix, that these tests were written for, has been reverted. So what would be the plan moving forward?

The original problem probably still remains in the SerialPort that the I/O stream is not implemented correctly and the read method doesn't block as defined by Java streams API

Patch-set: 1